fix(aero-machine): silence new clippy lints in tests
Remove identity arithmetic (addr + 0) and redundant u64 casts in newly added tests.

Keeps cargo clippy -p aero-machine --all-targets --all-features -- -D warnings clean.

+++ crates/aero-machine/tests/aerogpu_legacy_vbe_lfb_mmio.rs
@@ -52,7 +52,7 @@ fn aerogpu_legacy_vbe_lfb_banked_window_maps_into_vram() {
     let pitch = 64u64 * 4;
     let pixel_off = y * pitch + x * 4;
     let bank = (pixel_off / 65_536) as u16;
-    let bank_off = (pixel_off % 65_536) as u64;
+    let bank_off = pixel_off % 65_536;
 
     // Select bank 1 via Bochs VBE_DISPI.
     m.io_write(0x01CE, 2, 0x0005);
